+++ prjMyExamWeb1/frmHomeAdmin.aspx.cs
@@ -7,7 +7,7 @@ using System.Web.UI.WebControls;
 
 namespace prjMyExamWeb1
 {
-    public partial class WebForm1 : System.Web.UI.Page
+    public partial class frmHomeAdmin : System.Web.UI.Page
     {
         protected void Page_Load(object sender, EventArgs e)
         {
