+++ prjMyExamWeb1/prjMyExamWeb1.csproj.user
@@ -14,7 +14,7 @@
     <VisualStudio>
       <FlavorProperties GUID="{349c5851-65df-11da-9384-00065b846f21}">
         <WebProjectProperties>
-          <StartPageUrl>frmLogin1.aspx</StartPageUrl>
+          <StartPageUrl>frmLogin.aspx</StartPageUrl>
           <StartAction>SpecificPage</StartAction>
           <AspNetDebugging>True</AspNetDebugging>
           <SilverlightDebugging>False</SilverlightDebugging>
